Warning message when unfilled server settings

+++ Controller/EmailViewController.php
@@ -30,13 +30,19 @@ class EmailViewController extends BaseController
 
     public function load()
     {
+        $mailbox = $this->login();
         $task = $this->getTask();
-        $this->response->html($this->helper->layout->task('mailmagik:task_emails/task_load', array(
-            'task' => $task,
-            'project' => $this->projectModel->getById($task['project_id']),
-            'title'   => $task['title'],
-            'tags'    => $this->taskTagModel->getTagsByTask($task['id']),
-        )));
+        
+        if ($mailbox != false) {
+            $this->response->html($this->helper->layout->task('mailmagik:task_emails/task_load', array(
+                'task' => $task,
+                'project' => $this->projectModel->getById($task['project_id']),
+                'title'   => $task['title'],
+                'tags'    => $this->taskTagModel->getTagsByTask($task['id']),
+            )));
+        } else {
+            $this->response->redirect($this->helper->url->to('TaskViewController', 'show', array('task_id' => $task['id']), false, '', '', $this->request->isAjax()));   
+        }
     }
 
     public function view()
@@ -46,101 +52,103 @@ class EmailViewController extends BaseController
 
         $mailbox = $this->login();
 
-        try {
-            // Search in mailbox folder for specific emails
-            // PHP.net imap_search criteria: http://php.net/manual/en/function.imap-search.php
-            // Here, we search for "all" emails
-            $mails_ids = $mailbox->searchMailbox('TO ' . self::PREFIX);
-        } catch(PhpImap\Exceptions\ConnectionException $ex) {
-            die();
-        }
-
-        foreach ($mails_ids as $mail_id) {
-            $i = 0;
-
-            // Get mail by $mail_id
-            $email = $mailbox->getMail(
-                $mail_id, // ID of the email, you want to get
-                false // Do NOT mark emails as seen
-            );
-
-            $from_name = (isset($email->fromName)) ? $email->fromName : $email->fromAddress;
-            $from_email = $email->fromAddress;
-            foreach ($email->to as $to) {
-                if ($i === 0 && $to != null) {
-                    (strpos($to, self::PREFIX) == 0) ? $task_id = trim(str_replace(self::PREFIX, '', $to), ' ') : $task_id = null;
+        if ($mailbox != false) {
+            try {
+                // Search in mailbox folder for specific emails
+                // PHP.net imap_search criteria: http://php.net/manual/en/function.imap-search.php
+                // Here, we search for "all" emails
+                $mails_ids = $mailbox->searchMailbox('TO ' . self::PREFIX);
+            } catch(PhpImap\Exceptions\ConnectionException $ex) {
+                die();
+            }
+    
+            foreach ($mails_ids as $mail_id) {
+                $i = 0;
+    
+                // Get mail by $mail_id
+                $email = $mailbox->getMail(
+                    $mail_id, // ID of the email, you want to get
+                    false // Do NOT mark emails as seen
+                );
+    
+                $from_name = (isset($email->fromName)) ? $email->fromName : $email->fromAddress;
+                $from_email = $email->fromAddress;
+                foreach ($email->to as $to) {
+                    if ($i === 0 && $to != null) {
+                        (strpos($to, self::PREFIX) == 0) ? $task_id = trim(str_replace(self::PREFIX, '', $to), ' ') : $task_id = null;
+                    }
+                    $i++;
                 }
-                $i++;
-            }
-            $subject = $email->subject;
-            $message_id = $email->messageId;
-            $date = $email->date;
-
-            if ($email->textHtml) {
-                $email->embedImageAttachments();
-                $message = $email->textHtml;
-            } else {
-                $message = $email->textPlain;
-            }
-
-            if ($email->hasAttachments()) {
-                $has_attach = 'y';
-            } else {
-                $has_attach = 'n';
-            }
-
-            $attached_files = array();
-
-            $images = array();
-            if (!is_null($task_id) && intval($task_id) === intval($task['id'])) {
-                if (!empty($email->getAttachments())) {
-                    $attachments = $email->getAttachments();
-                    foreach ($attachments as $attachment) {
-                        $attached_files[] = $attachment->name;
-                        if (!file_exists(DATA_DIR . self::FILES_DIR . $task['id'])) {
-                            mkdir(DATA_DIR . self::FILES_DIR . $task['id'], 0755, true);
-                        }
-                        $attachment->setFilePath(DATA_DIR . self::FILES_DIR . $task['id'] . '/' . $attachment->name);
-                        if (!file_exists(DATA_DIR . self::FILES_DIR . $task['id'] . '/' . $attachment->name)) {
-                            $attachment->saveToDisk();
+                $subject = $email->subject;
+                $message_id = $email->messageId;
+                $date = $email->date;
+    
+                if ($email->textHtml) {
+                    $email->embedImageAttachments();
+                    $message = $email->textHtml;
+                } else {
+                    $message = $email->textPlain;
+                }
+    
+                if ($email->hasAttachments()) {
+                    $has_attach = 'y';
+                } else {
+                    $has_attach = 'n';
+                }
+    
+                $attached_files = array();
+    
+                $images = array();
+                if (!is_null($task_id) && intval($task_id) === intval($task['id'])) {
+                    if (!empty($email->getAttachments())) {
+                        $attachments = $email->getAttachments();
+                        foreach ($attachments as $attachment) {
+                            $attached_files[] = $attachment->name;
+                            if (!file_exists(DATA_DIR . self::FILES_DIR . $task['id'])) {
+                                mkdir(DATA_DIR . self::FILES_DIR . $task['id'], 0755, true);
+                            }
+                            $attachment->setFilePath(DATA_DIR . self::FILES_DIR . $task['id'] . '/' . $attachment->name);
+                            if (!file_exists(DATA_DIR . self::FILES_DIR . $task['id'] . '/' . $attachment->name)) {
+                                $attachment->saveToDisk();
+                            }
                         }
                     }
+    
+                    if (!$this->userModel->getByEmail($from_email)) {
+                        $connect_to_user = null;
+                    } else {
+                        $connect_to_user = $this->userModel->getByEmail($from_email);
+                    }
+    
+                    $emails[] = array(
+                        'mail_id' => $mail_id,
+                        'task_id' => $task_id,
+                        'project_id' => $this->projectModel->getById($task['project_id'])['id'],
+                        'from_name' => $from_name,
+                        'from_email' => $from_email,
+                        'subject' => $subject,
+                        'message_id' => $message_id,
+                        'message' => $message,
+                        'date' => $date,
+                        'has_attach' => $has_attach,
+                        'attachments' => $attached_files,
+                        'user' => $connect_to_user,
+                    );
                 }
-
-                if (!$this->userModel->getByEmail($from_email)) {
-                    $connect_to_user = null;
-                } else {
-                    $connect_to_user = $this->userModel->getByEmail($from_email);
-                }
-
-                $emails[] = array(
-                    'mail_id' => $mail_id,
-                    'task_id' => $task_id,
-                    'project_id' => $this->projectModel->getById($task['project_id'])['id'],
-                    'from_name' => $from_name,
-                    'from_email' => $from_email,
-                    'subject' => $subject,
-                    'message_id' => $message_id,
-                    'message' => $message,
-                    'date' => $date,
-                    'has_attach' => $has_attach,
-                    'attachments' => $attached_files,
-                    'user' => $connect_to_user,
-                );
+    
+                $mailbox->markMailAsRead($mail_id);
             }
-
-            $mailbox->markMailAsRead($mail_id);
+    
+            $emails = array_reverse($emails);
+    
+            $this->response->html($this->helper->layout->task('mailmagik:task_emails/task', array(
+                'task' => $task,
+                'project' => $this->projectModel->getById($task['project_id']),
+                'emails' => $emails,
+                'title'   => $task['title'],
+                'tags'    => $this->taskTagModel->getTagsByTask($task['id']),
+            )));
         }
-
-        $emails = array_reverse($emails);
-
-        $this->response->html($this->helper->layout->task('mailmagik:task_emails/task', array(
-            'task' => $task,
-            'project' => $this->projectModel->getById($task['project_id']),
-            'emails' => $emails,
-            'title'   => $task['title'],
-            'tags'    => $this->taskTagModel->getTagsByTask($task['id']),
-        )));
     }
 
     /**
@@ -202,51 +210,54 @@ class EmailViewController extends BaseController
         $task = $this->getTask();
 
         $mailbox = $this->login();
-        $email = $mailbox->getMail(
-            $mail_id,
-            false
-        );
-
-        $subject = $email->subject;
-        $message_id = $email->messageId;
-        $date = $email->date;
-
-        if ($email->textHtml) {
-            //$email->embedImageAttachments();
-            $message = $converter->convert($email->textHtml);
-        } else {
+        
+        if ($mailbox != false) {
+            $email = $mailbox->getMail(
+                $mail_id,
+                false
+            );
+    
+            $subject = $email->subject;
+            $message_id = $email->messageId;
+            $date = $email->date;
+    
+            if ($email->textHtml) {
+                //$email->embedImageAttachments();
+                $message = $converter->convert($email->textHtml);
+            } else {
+                $message = $email->textPlain;
+            }
             $message = $email->textPlain;
-        }
-        $message = $email->textPlain;
-
-
-        if ($email->hasAttachments()) {
-            $has_attach = 'y';
-        } else {
-            $has_attach = 'n';
-        }
-
-
-        $task_id = $this->taskCreationModel->create(array(
-            'project_id' => $task['project_id'],
-            'title' => $subject,
-            'description' => isset($message) ? $message : '',
-        ));
-
-        if (!empty($email->getAttachments())) {
-            $attachments = $email->getAttachments();
-            foreach ($attachments as $attachment) {
-                if (!file_exists(DATA_DIR . '/files/mailmagik/tmp/' . $task_id)) {
-                    mkdir(DATA_DIR . '/files/mailmagik/tmp/' . $task_id, 0755, true);
+    
+    
+            if ($email->hasAttachments()) {
+                $has_attach = 'y';
+            } else {
+                $has_attach = 'n';
+            }
+    
+    
+            $task_id = $this->taskCreationModel->create(array(
+                'project_id' => $task['project_id'],
+                'title' => $subject,
+                'description' => isset($message) ? $message : '',
+            ));
+    
+            if (!empty($email->getAttachments())) {
+                $attachments = $email->getAttachments();
+                foreach ($attachments as $attachment) {
+                    if (!file_exists(DATA_DIR . '/files/mailmagik/tmp/' . $task_id)) {
+                        mkdir(DATA_DIR . '/files/mailmagik/tmp/' . $task_id, 0755, true);
+                    }
+                    $tmp_name = DATA_DIR . '/files/mailmagik/tmp/' . $task_id . '/' . $attachment->name;
+                    $attachment->setFilePath($tmp_name);
+                    if (!file_exists($tmp_name)) {
+                        $attachment->saveToDisk();
+                    }
+                    $file = file_get_contents($tmp_name);
+                    $this->taskFileModel->uploadContent($task_id, $attachment->name, $file, false);
+                    unlink($tmp_name);
                 }
-                $tmp_name = DATA_DIR . '/files/mailmagik/tmp/' . $task_id . '/' . $attachment->name;
-                $attachment->setFilePath($tmp_name);
-                if (!file_exists($tmp_name)) {
-                    $attachment->saveToDisk();
-                }
-                $file = file_get_contents($tmp_name);
-                $this->taskFileModel->uploadContent($task_id, $attachment->name, $file, false);
-                unlink($tmp_name);
             }
         }
 
@@ -275,59 +286,63 @@ class EmailViewController extends BaseController
         $task_id =  $this->request->getIntegerParam('task_id');
         $task = $this->getTask();
         $mailbox = $this->login();
-        $email = $mailbox->getMail(
-            $mail_id,
-            false
-        );
-
-        $subject = $email->subject;
-        $message_id = $email->messageId;
-        $date = $email->date;
-
-        if ($email->textHtml) {
-            //$email->embedImageAttachments();
-            $message = $converter->convert($email->textHtml);
-        } else {
+        
+        if ($mailbox != false) {
+            $email = $mailbox->getMail(
+                $mail_id,
+                false
+            );
+    
+            $subject = $email->subject;
+            $message_id = $email->messageId;
+            $date = $email->date;
+    
+            if ($email->textHtml) {
+                //$email->embedImageAttachments();
+                $message = $converter->convert($email->textHtml);
+            } else {
+                $message = $email->textPlain;
+            }
             $message = $email->textPlain;
+            $from_email = $email->fromAddress;
+    
+            if ($email->hasAttachments()) {
+                $has_attach = 'y';
+            } else {
+                $has_attach = 'n';
+            }
+    
+            if (!$this->userModel->getByEmail($from_email)) {
+                $connect_to_user = null;
+                $comment = '*Email converted to comment and originally sent by ' . $from_email . '*' ."\n\n" . (isset($subject) ? "#$subject\n\n" : '') . (isset($message) ? $message : '');
+            } else {
+                $connect_to_user = $this->userModel->getByEmail($from_email);
+                $comment = '*Email converted to comment by ' . $user['username']. '*' ."\n\n" . (isset($subject) ? "#$subject\n\n" : '') . (isset($message) ? $message : '');
+            }
+    
+            $values = array(
+                'task_id' => $task_id,
+                'comment' => $comment,
+                'user_id' => is_null($connect_to_user) ? $user['id'] : $connect_to_user['id'],
+            );
+    
+    
+            $comment_id = $this->commentModel->create($values);
+    
+    
+            $option = $this->configModel->get('mailmagik_pref', '2');
+    
+            if ($option == 2) {
+                $mailbox->markMailAsRead($mail_id);
+            } else {
+                $mailbox->deleteMail($mail_id);
+            }
+    
+            $task_id = $task_id.'#comment-'.$comment_id;
+    
+            $this->response->redirect($this->helper->url->to('TaskViewController', 'show', array('task_id' => $task_id), false, '', '', $this->request->isAjax(), 'comment-'.$comment_id));
+            
         }
-        $message = $email->textPlain;
-        $from_email = $email->fromAddress;
-
-        if ($email->hasAttachments()) {
-            $has_attach = 'y';
-        } else {
-            $has_attach = 'n';
-        }
-
-        if (!$this->userModel->getByEmail($from_email)) {
-            $connect_to_user = null;
-            $comment = '*Email converted to comment and originally sent by ' . $from_email . '*' ."\n\n" . (isset($subject) ? "#$subject\n\n" : '') . (isset($message) ? $message : '');
-        } else {
-            $connect_to_user = $this->userModel->getByEmail($from_email);
-            $comment = '*Email converted to comment by ' . $user['username']. '*' ."\n\n" . (isset($subject) ? "#$subject\n\n" : '') . (isset($message) ? $message : '');
-        }
-
-        $values = array(
-            'task_id' => $task_id,
-            'comment' => $comment,
-            'user_id' => is_null($connect_to_user) ? $user['id'] : $connect_to_user['id'],
-        );
-
-
-        $comment_id = $this->commentModel->create($values);
-
-
-        $option = $this->configModel->get('mailmagik_pref', '2');
-
-        if ($option == 2) {
-            $mailbox->markMailAsRead($mail_id);
-        } else {
-            $mailbox->deleteMail($mail_id);
-        }
-
-        $task_id = $task_id.'#comment-'.$comment_id;
-
-        $this->response->redirect($this->helper->url->to('TaskViewController', 'show', array('task_id' => $task_id), false, '', '', $this->request->isAjax(), 'comment-'.$comment_id));
     }
 
 
@@ -342,14 +357,21 @@ class EmailViewController extends BaseController
         $port = $this->configModel->get('mailmagik_port', '');
         $user = $this->configModel->get('mailmagik_user', '');
         $password = $this->configModel->get('mailmagik_password', '');
-
-        $mailbox = new PhpImap\Mailbox(
-            '{'.$server.':' . $port . '/imap/ssl}INBOX',
-            $user,
-            $password,
-            false
-        );
-
-        return $mailbox;
+        
+        if ($server != '' && $port != '' && $user != '' && $password != '') {
+            
+            $mailbox = new PhpImap\Mailbox(
+                '{'.$server.':' . $port . '/imap/ssl}INBOX',
+                $user,
+                $password,
+                false
+            );
+    
+            return $mailbox;
+            
+        } else {
+            $this->flash->failure(t('IMAP Server is missing config settings information!'));
+            return false;
+        }
     }
 }
